+++ Sources/Sentry/SentryTimeToDisplayTracker.m
@@ -2,7 +2,6 @@
 
 #if SENTRY_HAS_UIKIT
 
-#    import "SentryCurrentDateProvider.h"
 #    import "SentryDependencyContainer.h"
 #    import "SentryFramesTracker.h"
 #    import "SentryMeasurementValue.h"
@@ -99,15 +98,13 @@ SentryTimeToDisplayTracker () <SentryFramesTrackerListener>
     _fullyDisplayedReported = YES;
 }
 
-- (void)framesTrackerHasNewFrame
+- (void)framesTrackerHasNewFrame:(NSDate *)newFrameDate
 {
-    NSDate *finishTime = [SentryDependencyContainer.sharedInstance.dateProvider date];
-
     // The purpose of TTID and TTFD is to measure how long
     // takes to the content of the screen to change.
     // Thats why we need to wait for the next frame to be drawn.
     if (_initialDisplayReported && self.initialDisplaySpan.isFinished == NO) {
-        self.initialDisplaySpan.timestamp = finishTime;
+        self.initialDisplaySpan.timestamp = newFrameDate;
 
         [self.initialDisplaySpan finish];
 
@@ -117,7 +114,7 @@ SentryTimeToDisplayTracker () <SentryFramesTrackerListener>
     }
     if (_waitForFullDisplay && _fullyDisplayedReported && self.fullDisplaySpan.isFinished == NO
         && self.initialDisplaySpan.isFinished == YES) {
-        self.fullDisplaySpan.timestamp = finishTime;
+        self.fullDisplaySpan.timestamp = newFrameDate;
 
         [self.fullDisplaySpan finish];
     }